+++ driver-kafka/build.gradle
@@ -10,7 +10,7 @@ repositories {
 }
 
 dependencies {
-    compile project(":dsb-api")
+    compile project(":sbk-api")
     compile	"org.apache.kafka:kafka-clients:2.3.0"
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
