+++ driver-pravega/build.gradle
@@ -10,7 +10,7 @@ repositories {
 }
 
 dependencies {
-    compile project(":dsb-api")
+    compile project(":sbk-api")
     compile "io.pravega:pravega-client:0.6.0"
     compile	"io.pravega:pravega-common:0.6.0"
     testCompile group: 'junit', name: 'junit', version: '4.12'
