+++ driver-pulsar/build.gradle
@@ -10,7 +10,7 @@ repositories {
 }
 
 dependencies {
-    compile project(":dsb-api")
+    compile project(":sbk-api")
     compile	"org.apache.pulsar:pulsar-client:2.4.2"
     compile	"org.apache.pulsar:pulsar-client-admin:2.4.2"
     testCompile group: 'junit', name: 'junit', version: '4.12'
