--- driver-pulsar/src/main/java/io/dsb/Pulsar/PulsarReader.java
@@ -1,68 +1,0 @@
-/**
- * Copyright (c) 2020 KMG. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- */
-package io.dsb.Pulsar;
-
-import io.dsb.api.Reader;
-import io.dsb.api.QuadConsumer;
-import io.dsb.api.Parameters;
-
-import java.io.IOException;
-import java.util.concurrent.TimeUnit;
-
-import org.apache.pulsar.client.api.PulsarClient;
-import org.apache.pulsar.client.api.Consumer;
-import org.apache.pulsar.client.api.SubscriptionType;
-import org.apache.pulsar.client.api.PulsarClientException;
-import org.apache.pulsar.client.api.SubscriptionInitialPosition;
-
-/**
- * Class for Pulsar reader/consumer.
- */
-public class PulsarReader extends Reader {
-    final private Consumer<byte[]> consumer;
-
-    public PulsarReader(int readerId, Parameters params, QuadConsumer recordTime,
-                              String topicName, String subscriptionName, PulsarClient client) throws  IOException {
-        super(readerId, params, recordTime);
-        final  SubscriptionInitialPosition position = params.writeAndRead ? SubscriptionInitialPosition.Latest :
-                                                    SubscriptionInitialPosition.Earliest;
-        try {
-            this.consumer = client.newConsumer()
-                    .topic(topicName)
-                    // Allow multiple consumers to attach to the same subscription
-                    // and get messages dispatched as a queue
-                    .subscriptionType(SubscriptionType.Exclusive)
-                    .subscriptionName(subscriptionName)
-                    .subscriptionInitialPosition(position)
-                    .receiverQueueSize(1)
-                    .subscribe();
-        } catch (PulsarClientException ex) {
-            throw new IOException(ex);
-        }
-    }
-
-    @Override
-    public byte[] read() throws IOException {
-        try {
-            return consumer.receive(params.timeout, TimeUnit.SECONDS).getData();
-        } catch (PulsarClientException ex) {
-            throw new IOException(ex);
-        }
-    }
-
-    @Override
-    public void close() throws IOException {
-        try {
-            consumer.close();
-        } catch (PulsarClientException ex) {
-            throw new IOException(ex);
-        }
-    }
-}
